Some basic Refactoring
